+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 - Calculation wallet balance
+- Database: add Wallet entity
 - RPC requests: getWalletBalance
 
 ## [0.2.0] - 2018-06-29